+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log  
 
+## [2017-08-11]  
++ **Added**  
+	- `Best-Time-to-Buy-and-Sell-Stock-II` for [problem](https://leetcode.com/problems/best-time-to-buy-and-sell-stock-ii/description/)     
+	- `Best-Time-to-Buy-and-Sell-Stock-with-Cooldown` for [problem](https://leetcode.com/problems/best-time-to-buy-and-sell-stock-with-cooldown/description/)   
+	- `Best-Time-to-Buy-and-Sell-Stock` for [problem](https://leetcode.com/problems/best-time-to-buy-and-sell-stock/description/)  
+	- `Range-Sum-Query-Immutable` for [problem](https://leetcode.com/problems/range-sum-query-immutable/description/)  
+
 ## [2017-08-08]  
 + **Added**  
 	- `Pow(x,n)` implements an operation calculating the `x` to the power of `n`